added translateUsage function and formatted booking dialog content
