+++ common/translateUsage.ts
@@ -1,0 +1,24 @@
+const translateUsage = ( usage : string  ) => {
+        switch(usage) {
+            case "AUD" :
+                return "Auditorium";
+            case "CMLB":
+                return "Computer Lab";
+            case "LAB" : 
+                return "Lab";
+            case "LCTR":
+                return "Lecture Hall";
+            case "MEET":
+                return "Meeting Room";
+            case "SDIO":
+               return "Studio";
+            case "TUSM":
+                return "Tutorial Room";
+            case "LIB":
+                return "Library";
+            default:
+                return "";
+        }
+    }
+
+export default translateUsage;
